requirements.txt
requests>=2.26.0
plyer>=2.0.0
pystray>=0.19.2
pillow>=9.0.0
watchdog>=2.1.6
python-dotenv>=0.19.0
pywin32>=300; sys_platform == 'win32'